rename: #82 배송담당자 배정 서비스 디렉토리 변경
